+++ xarray/backends/pynio_.py
@@ -7,39 +7,43 @@ import functools
 import numpy as np
 
 from .. import Variable
-from ..core.utils import FrozenOrderedDict, Frozen, NDArrayMixin
+from ..core.utils import (FrozenOrderedDict, Frozen,
+                          NdimSizeLenMixin, DunderArrayMixin)
 from ..core import indexing
 
 from .common import AbstractDataStore, DataStorePickleMixin
 
 
-class NioArrayWrapper(NDArrayMixin):
+class NioArrayWrapper(NdimSizeLenMixin, DunderArrayMixin):
 
     def __init__(self, variable_name, datastore):
         self.datastore = datastore
         self.variable_name = variable_name
+        array = self.get_array()
+        self.shape = array.shape
+        self.dtype = np.dtype(array.typecode())
 
-    @property
-    def array(self):
+    def get_array(self):
+        self.datastore.assert_open()
         return self.datastore.ds.variables[self.variable_name]
 
-    @property
-    def dtype(self):
-        return np.dtype(self.array.typecode())
-
     def __getitem__(self, key):
-        if key == () and self.ndim == 0:
-            return self.array.get_value()
-        return self.array[key]
+        with self.datastore.ensure_open(autoclose=True):
+            array = self.get_array()
+            if key == () and self.ndim == 0:
+                return array.get_value()
+            return array[key]
 
 
 class NioDataStore(AbstractDataStore, DataStorePickleMixin):
     """Store for accessing datasets via PyNIO
     """
-    def __init__(self, filename, mode='r'):
+    def __init__(self, filename, mode='r', autoclose=False):
         import Nio
         opener = functools.partial(Nio.open_file, filename, mode=mode)
         self.ds = opener()
+        self._autoclose = autoclose
+        self._isopen = True
         self._opener = opener
         self._mode = mode
 
@@ -48,14 +52,17 @@ class NioDataStore(AbstractDataStore, DataStorePickleMixin):
         return Variable(var.dimensions, data, var.attributes)
 
     def get_variables(self):
-        return FrozenOrderedDict((k, self.open_store_variable(k, v))
-                                 for k, v in self.ds.variables.iteritems())
+        with self.ensure_open(autoclose=False):
+            return FrozenOrderedDict((k, self.open_store_variable(k, v))
+                                     for k, v in self.ds.variables.iteritems())
 
     def get_attrs(self):
-        return Frozen(self.ds.attributes)
+        with self.ensure_open(autoclose=True):
+            return Frozen(self.ds.attributes)
 
     def get_dimensions(self):
-        return Frozen(self.ds.dimensions)
+        with self.ensure_open(autoclose=True):
+            return Frozen(self.ds.dimensions)
 
     def get_encoding(self):
         encoding = {}
@@ -64,4 +71,6 @@ class NioDataStore(AbstractDataStore, DataStorePickleMixin):
         return encoding
 
     def close(self):
-        self.ds.close()
+        if self._isopen:
+            self.ds.close()
+            self._isopen = False